Move get game actions into slice logic file
