+++ src/slices/game/game-slice.ts
@@ -1,12 +1,21 @@
-import { createAsyncThunk, createSlice, PayloadAction, Reducer } from '@reduxjs/toolkit';
+import { Action, ActionCreator, createAsyncThunk, createSlice, PayloadAction, Reducer } from '@reduxjs/toolkit';
 import { Game, GameDetail, Games, GameStatus } from '../../models/game.js';
+import { Roster } from '../../models/player.js';
 import { oldReducer } from '../../reducers/game.js';
-import { RootState, ThunkResult } from '../../store.js';
+import {
+  GET_GAME_FAIL, GET_GAME_REQUEST,
+  GET_GAME_SUCCESS
+} from '../../slices/game-types.js';
+import { RootState, ThunkPromise, ThunkResult } from '../../store.js';
 import { selectCurrentUserId } from '../auth/auth-slice.js';
 import { gameCompleted, gameSetupCompleted, selectLiveGameById } from '../live/live-slice.js';
-import { loadGames, persistNewGame, updateExistingGame } from './game-storage.js';
+import { loadGame, loadGameRoster, loadGames, persistNewGame, updateExistingGame } from './game-storage.js';
 import { copyRoster, rosterCopiedHandler, rosterCopyFailedHandler, rosterCopyPendingHandler } from './roster-logic.js';
 export { addNewGamePlayer, copyRoster } from './roster-logic.js';
+
+interface GameActionGetGameRequest extends Action<typeof GET_GAME_REQUEST> { gameId: string };
+export interface GameActionGetGameSuccess extends Action<typeof GET_GAME_SUCCESS> { game: GameDetail };
+interface GameActionGetGameFail extends Action<typeof GET_GAME_FAIL> { error: string };
 
 export const getGames = createAsyncThunk<
   // Return type of the payload creator
@@ -34,6 +43,66 @@ export const getGames = createAsyncThunk<
     },
   }
 );
+
+export const getGame = (gameId: string): ThunkPromise<void> => (dispatch, getState) => {
+  if (!gameId) {
+    return Promise.reject();
+  }
+  dispatch(getGameRequest(gameId));
+
+  // TODO: Check for game cached in IDB (similar to getTeams action)
+  // Checks if the game has already been retrieved.
+  const state = getState();
+  let existingGame: Game | undefined;
+  if (state.game && state.game.game && state.game.game.id === gameId) {
+    existingGame = state.game.game!;
+  } else {
+    existingGame = state.game?.games[gameId];
+  }
+  if (existingGame && existingGame.hasDetail) {
+    dispatch(getGameSuccess(existingGame));
+    // let the calling code know there's nothing to wait for.
+    return Promise.resolve();
+  }
+
+  let game: Game;
+  return loadGame(gameId)
+    .then((result) => {
+      game = result;
+      return loadGameRoster(gameId);
+    })
+    .then((gameRoster: Roster) => {
+      const gameDetail: GameDetail = {
+        ...game,
+        roster: gameRoster
+      };
+      dispatch(getGameSuccess(gameDetail));
+    })
+    .catch((error: any) => {
+      dispatch(getGameFail(error.toString()));
+    });
+}
+
+const getGameRequest: ActionCreator<GameActionGetGameRequest> = (gameId: string) => {
+  return {
+    type: GET_GAME_REQUEST,
+    gameId
+  };
+};
+
+const getGameSuccess: ActionCreator<GameActionGetGameSuccess> = (game: GameDetail) => {
+  return {
+    type: GET_GAME_SUCCESS,
+    game
+  };
+};
+
+const getGameFail: ActionCreator<GameActionGetGameFail> = (error: string) => {
+  return {
+    type: GET_GAME_FAIL,
+    error
+  };
+};
 
 export const addNewGame = (newGame: Game): ThunkResult => (dispatch, getState) => {
   if (!newGame) {
